flask/requirements.txt
Flask==2.0.1
psycopg2-binary==2.9.1
redis==3.5.3
Werkzeug==2.2.2

